backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.1
cryptography==42.0.5
PyJWT==2.8.0
werkzeug==3.0.1
